# 02-development/homework/balancio/openapi.yaml
openapi: 3.0.3
info:
  title: Balancio API
  description: >
    Shared expense splitter for friends, households, couples, trips, and
    other small groups. See _docs/specs.md for the full product spec.
  version: "0.1.0"
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /groups:
    get:
      tags: [groups]
      summary: List groups (admin sees all, others see their own)
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      tags: [groups]
      summary: Create a group (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /groups/{group_id}:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    get:
      tags: [groups]
      summary: Get a group by id
      responses:
        "200":
          description: The group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [groups]
      summary: Update a group's name/description (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/members:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    post:
      tags: [groups]
      summary: Add a member to a group (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Group is archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/members/{user_id}:
    delete:
      tags: [groups]
      summary: Remove a member from a group (admin only, requires zero group balance)
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Member has a non-zero balance in this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/leave:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    post:
      tags: [groups]
      summary: Leave a group (requires zero balance in that group)
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Balance in this group is not zero
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/archive:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    post:
      tags: [groups]
      summary: Archive a group (admin only, requires all balances zero)
      responses:
        "200":
          description: Archived group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Group has non-zero balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/balances:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    get:
      tags: [balances]
      summary: Get the expense-derived balance breakdown for a group
      responses:
        "200":
          description: Group balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupBalances"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /refunds:
    get:
      tags: [refunds]
      summary: List refunds
      responses:
        "200":
          description: List of refunds, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Refund"
    post:
      tags: [refunds]
      summary: >
        Create a refund. The creator's own confirmation is recorded
        immediately; balances are unaffected until every participant confirms.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRefundRequest"
      responses:
        "201":
          description: Refund created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /refunds/{refund_id}:
    parameters:
      - $ref: "#/components/parameters/RefundId"
    get:
      tags: [refunds]
      summary: Get a refund by id
      responses:
        "200":
          description: The refund
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
        "404":
          $ref: "#/components/responses/NotFound"
  /refunds/{refund_id}/confirm:
    parameters:
      - $ref: "#/components/parameters/RefundId"
    post:
      tags: [refunds]
      summary: >
        Confirm a refund as the current user. Once every participant has
        confirmed, the refund transitions to confirmed and affects balances.
      responses:
        "200":
          description: Updated refund
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
        "400":
          description: Not a participant, or already confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
...
    RefundId:
      name: refund_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation error or invalid financial state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Friendly, domain-specific message. Never a stack trace.
...
    Money:
      type: string
      description: Decimal amount with exactly two decimal places.
      pattern: '^\d+\.\d{2}$'
      example: "123.45"
...
    GroupStatus:
      type: string
      enum: [active, archived]
    Group:
      type: object
      required: [id, name, member_ids, status, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        member_ids:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/GroupStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name, member_ids]
      properties:
        name:
          type: string
        description:
          type: string
        member_ids:
          type: array
          items:
            type: string
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    RefundStatus:
      type: string
      enum: [pending, confirmed]
    RefundConfirmation:
      type: object
      required: [user_id, confirmed]
      properties:
        user_id:
          type: string
        confirmed:
          type: boolean
    Refund:
      type: object
      required:
        - id
        - group_id
        - title
        - amount
        - participant_ids
        - created_by
        - status
        - confirmations
        - created_at
        - updated_at
      properties:
        id:
          type: string
        group_id:
          type: string
        title:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        participant_ids:
          type: array
          items:
            type: string
        created_by:
          type: string
        status:
          $ref: "#/components/schemas/RefundStatus"
        confirmations:
          type: array
          items:
            $ref: "#/components/schemas/RefundConfirmation"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRefundRequest:
      type: object
      required: [group_id, title, amount, participant_ids]
      properties:
        group_id:
          type: string
        title:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        participant_ids:
          type: array
          items:
            type: string
...
    PairBalance:
      type: object
      required: [from_user_id, to_user_id, amount]
      properties:
        from_user_id:
          type: string
          description: The user who owes.
        to_user_id:
          type: string
          description: The user who is owed.
        amount:
          $ref: "#/components/schemas/Money"
...
    GroupBalances:
      type: object
      required: [group_id, net, pairs]
      properties:
        group_id:
          type: string
        net:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Money"
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/PairBalance"
...